Move components into classified subdirs Update rest of core
