+++ core/components/scriptable/PointLightComponent.cc
@@ -1,7 +1,7 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/components/PointLightComponent.h"
+#include "core/components/scriptable/PointLightComponent.h"
 
 #include "types/protocol/WorldEvent_generated.h"
