+++ core/components/synchronized/MeshRendererComponent.cc
@@ -1,7 +1,7 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/components/MeshRendererComponent.h"
+#include "core/components/synchronized/MeshRendererComponent.h"
 
 #include "types/protocol/WorldEvent_generated.h"
