+++ core/components/synchronized/MeshRendererComponent.h
@@ -5,7 +5,7 @@
 
 #include "core/assets/MaterialAsset.h"
 #include "core/assets/MeshAsset.h"
-#include "core/components/Component.h"
+#include "core/components/SynchronizedComponent.h"
 #include "types/assets/PrefabAsset_generated.h"
 #include "types/protocol/MeshRendererComponent_generated.h"
